// src/data/enums.ts
export enum HomeFeedType {
  COLLABORATIONS = "COLLABORATIONS",
  FOLLOWING = "FOLLOWING",
  FORYOU = "FORYOU",
  HIGHLIGHTS = "HIGHLIGHTS",
  SPORTS = "SPORTS",
  LIFESTYLE = "LIFESTYLE",
  POP_CULTURE = "POP_CULTURE",
  PODCASTS = "PODCASTS",
  PHOTOGRAPHY = "PHOTOGRAPHY",
  FOOD = "FOOD",
  WRITERS = "WRITERS",
  COMMUNITIES = "COMMUNITIES",
  COMEDIANS = "COMEDIANS"
}

export enum HomeFeedView {
  LIST = "LIST",
  GRID = "GRID"
}

export enum HomeFeedSort {
  LATEST = "LATEST",
  OLDEST = "OLDEST"
}

export enum AccountFeedType {
  Collaborations = "COLLABORATIONS",
  Collects = "COLLECTS",
  FanDrops = "FANDROPS",
  Feed = "FEED",
  Media = "MEDIA",
  Replies = "REPLIES"
}

export enum NotificationFeedType {
  All = "ALL",
  Activity = "ACTIVITY",
  Referrals = "REFERRALS",
  Rewards = "REWARDS",
  System = "SYSTEM"
}

export enum GroupsFeedType {
  Discover = "DISCOVER",
  Managed = "MANAGED",
  Member = "MEMBER"
}

export enum Status {
  Success = "success",
  Error = "error"
}